changes to get it working on mas-support
I've used `uv` to set up a pyproject.toml file, and used `nvm` for the
frontend. Hopefully we'll have a reproducible environment.

I've removed requirements-base.txt and requirements-unix.txt - the
pyproject.toml should be the single source of dependencies.
requirements-base.txt seemed to just be a dump of `pip freeze`; I think
I've whittled it down to just the packages that need to be directly
installed.

I seem to have ended up with different versions of some packages, which
meant that I had to make some changes.

The musical_scales.scales function seems to have made some arguments
keyword-only.

The sonify_lightcurve endpoint should return HTTP code 500 if there's an
error, not 404.

When building the frontend, it looks for an environment variable
VITE_SONIFICATION_BACKEND_URL, instead of hardcoding
"http://localhost:8000". In production, this should be "/api/v1"

There were some references to `lightCurvesAPI` which I think are
supposed to be `apiUrl`.

`src/frontend/src/components/ui/Buttons.tsx` was trying to import a name
`Variant` from `pages/Lightcurves`, but that doesn't exist, and it isn't
used, so I removed the import.